Updated target framework and FV version. Refactored to eliminate warnings.
