+++ FluentValidationLister.Filter/Internal/StringExtensions.cs
@@ -2,64 +2,63 @@
 // Copyright (c) Dan Ware. All rights reserved.
 // </copyright>
 
-namespace FluentValidationLister.Filter.Internal
+namespace FluentValidationLister.Filter.Internal;
+
+using System.Collections.Generic;
+using System.Globalization;
+
+/// <summary>
+/// Class to provide additional functionality for strings.
+/// </summary>
+internal static class StringExtensions
 {
-    using System.Collections.Generic;
-    using System.Globalization;
-
     /// <summary>
-    /// Class to provide additional functionality for strings.
+    /// Returns the string with the first character switched to lower case.
     /// </summary>
-    internal static class StringExtensions
+    /// <param name="input">Current string for which to make the first letter lower case.</param>
+    /// <returns>Formatted string.</returns>
+    public static string ToCamelCase(this string input)
     {
-        /// <summary>
-        /// Returns the string with the first character switched to lower case.
-        /// </summary>
-        /// <param name="input">Current string for which to make the first letter lower case.</param>
-        /// <returns>Formatted string.</returns>
-        public static string ToCamelCase(this string input)
+        if (string.IsNullOrEmpty(input) || !char.IsUpper(input[0]))
         {
-            if (string.IsNullOrEmpty(input) || !char.IsUpper(input[0]))
-            {
-                return input;
-            }
-
-            if (input.Contains('.'))
-            {
-                var parts = input.Split('.');
-                var camelCasedParts = new List<string>();
-                foreach (var part in parts)
-                {
-                    camelCasedParts.Add(StringToCamelCase(part));
-                }
-
-                return string.Join(".", camelCasedParts);
-            }
-
-            return StringToCamelCase(input);
+            return input;
         }
 
-        private static string StringToCamelCase(string input)
+        if (input.Contains('.'))
         {
-            var chars = input.ToCharArray();
-
-            for (var i = 0; i < chars.Length; i++)
+            var parts = input.Split('.');
+            var camelCasedParts = new List<string>();
+            foreach (var part in parts)
             {
-                if (i == 1 && !char.IsUpper(chars[i]))
-                {
-                    break;
-                }
-
-                var hasNext = i + 1 < chars.Length;
-                if (i > 0 && hasNext && !char.IsUpper(chars[i + 1]))
-                {
-                    break;
-                }
-
-                chars[i] = char.ToLower(chars[i], CultureInfo.InvariantCulture);
+                camelCasedParts.Add(StringToCamelCase(part));
             }
 
-            return new string(chars);
+            return string.Join(".", camelCasedParts);
         }
+
+        return StringToCamelCase(input);
+    }
+
+    private static string StringToCamelCase(string input)
+    {
+        var chars = input.ToCharArray();
+
+        for (var i = 0; i < chars.Length; i++)
+        {
+            if (i == 1 && !char.IsUpper(chars[i]))
+            {
+                break;
+            }
+
+            var hasNext = i + 1 < chars.Length;
+            if (i > 0 && hasNext && !char.IsUpper(chars[i + 1]))
+            {
+                break;
+            }
+
+            chars[i] = char.ToLower(chars[i], CultureInfo.InvariantCulture);
+        }
+
+        return new string(chars);
     }
 }
