+++ FluentValidationLister.Filter/Meta/FromToValues.cs
@@ -2,38 +2,37 @@
 // Copyright (c) Dan Ware. All rights reserved.
 // </copyright>
 
-namespace FluentValidationLister.Filter.Meta
+namespace FluentValidationLister.Filter.Meta;
+
+using System;
+
+/// <summary>
+/// Class to hold serialisable properties for From and To.
+/// </summary>
+public class FromToValues
 {
-    using System;
+    /// <summary>
+    /// Initialises a new instance of the <see cref="FromToValues"/> class.
+    /// </summary>
+    public FromToValues()
+    {
+    }
 
     /// <summary>
-    /// Class to hold serialisable properties for From and To.
+    /// Initialises a new instance of the <see cref="FromToValues"/> class with specified
+    /// values for the From and To properties.
     /// </summary>
-    public class FromToValues
+    /// <param name="from">Value for the From property.</param>
+    /// <param name="to">Value for the To property.</param>
+    public FromToValues(IComparable from, IComparable to)
     {
-        /// <summary>
-        /// Initialises a new instance of the <see cref="FromToValues"/> class.
-        /// </summary>
-        public FromToValues()
-        {
-        }
-
-        /// <summary>
-        /// Initialises a new instance of the <see cref="FromToValues"/> class with specified
-        /// values for the From and To properties.
-        /// </summary>
-        /// <param name="from">Value for the From property.</param>
-        /// <param name="to">Value for the To property.</param>
-        public FromToValues(IComparable from, IComparable to)
-        {
-            this.From = from is int @int ? @int : throw new InvalidOperationException($"Cannot set range for type {from.GetType()}");
-            this.To = to is int int1 ? int1 : throw new InvalidOperationException($"Cannot set range for type {to.GetType()}");
-        }
-
-        /// <summary>Gets or sets the From value.</summary>
-        public int From { get; set; }
-
-        /// <summary>Gets or sets the To value.</summary>
-        public int To { get; set; }
+        this.From = from is int @int ? @int : throw new InvalidOperationException($"Cannot set range for type {from.GetType()}");
+        this.To = to is int int1 ? int1 : throw new InvalidOperationException($"Cannot set range for type {to.GetType()}");
     }
+
+    /// <summary>Gets or sets the From value.</summary>
+    public int From { get; set; }
+
+    /// <summary>Gets or sets the To value.</summary>
+    public int To { get; set; }
 }
