+++ FluentValidationLister.Tests/Models/Country.cs
@@ -1,0 +1,12 @@
+﻿// <copyright file="Country.cs" company="Dan Ware">
+// Copyright (c) Dan Ware. All rights reserved.
+// </copyright>
+
+namespace FluentValidationLister.Tests.Models;
+
+/// <summary> Country entity object. </summary>
+public class Country
+{
+    /// <summary> Gets or sets the name. </summary>
+    public string Name { get; set; }
+}
