+++ FluentValidationLister.Tests/Models/EnumGender.cs
@@ -1,0 +1,15 @@
+﻿// <copyright file="EnumGender.cs" company="Dan Ware">
+// Copyright (c) Dan Ware. All rights reserved.
+// </copyright>
+
+namespace FluentValidationLister.Tests.Models;
+
+/// <summary> Gender enumerable. </summary>
+public enum EnumGender
+{
+    /// <summary> Female gender. </summary>
+    Female = 1,
+
+    /// <summary> Male gender. </summary>
+    Male = 2,
+}
