--- FluentValidationLister.Tests/Samples/Person.cs
@@ -1,124 +1,0 @@
-﻿namespace FluentValidationLister.Tests.Samples
-{
-    using System;
-    using System.Collections.Generic;
-
-    /// <summary>
-    /// Dummy entity for testing purposes.
-    /// </summary>
-    /// <remarks><see href="https://github.com/JeremySkinner/FluentValidation/blob/master/src/FluentValidation.Tests/Person.cs">Original source.</see></remarks>
-    public class Person
-    {
-        public int Id { get; set; }
-
-        public string Surname { get; set; }
-
-        public string Forename { get; set; }
-
-        public bool IsActive { get; set; } = true;
-
-        public List<Person> Children { get; set; }
-
-        public DateTime DateOfBirth { get; set; }
-
-        public int? NullableInt { get; set; }
-
-        public Person()
-        {
-            Children = new List<Person>();
-            Orders = new List<Order>();
-        }
-
-        public static int CalculateSalary()
-        {
-            return 20;
-        }
-
-        public Address Address { get; set; }
-
-        public IList<Order> Orders { get; set; }
-
-        public string Email { get; set; }
-
-        public decimal? Discount { get; set; }
-
-        public double? Age { get; set; }
-
-        public int AnotherInt { get; set; }
-
-        public string CreditCard { get; set; }
-
-        public int? OtherNullableInt { get; set; }
-
-        public string Regex { get; set; }
-
-        public System.Text.RegularExpressions.Regex AnotherRegex { get; set; }
-
-        public int MinLength { get; set; }
-
-        public int MaxLength { get; set; }
-
-        public EnumGender Gender { get; set; }
-
-        public string GenderString { get; set; }
-
-        public string ForenameReadOnly => Forename;
-
-        public float? SomeSingleValue { get; set; } = (float)1.0;
-    }
-
-    public interface IAddress
-    {
-        string Line1 { get; set; }
-
-        string Line2 { get; set; }
-
-        string Town { get; set; }
-
-        string County { get; set; }
-
-        string Postcode { get; set; }
-
-        Country Country { get; set; }
-    }
-
-    public class Address : IAddress
-    {
-        public string Line1 { get; set; }
-
-        public string Line2 { get; set; }
-
-        public string Town { get; set; }
-
-        public string County { get; set; }
-
-        public string Postcode { get; set; }
-
-        public Country Country { get; set; }
-
-        public int Id { get; set; }
-    }
-
-    public class Country
-    {
-        public string Name { get; set; }
-    }
-
-    public interface IOrder
-    {
-        decimal Amount { get; }
-    }
-
-    public class Order : IOrder
-    {
-        public string ProductName { get; set; }
-
-        public decimal Amount { get; set; }
-    }
-
-    public enum EnumGender
-    {
-        Female = 1,
-        Male = 2,
-    }
-}
