+++ FluentValidationLister.WebApiSample/Scripts/test-validation.ts
@@ -118,16 +118,16 @@ $(function () {
                         }
                         break;
                     case "lessThan":
-                        if (fieldHasValue === true) fieldPassesRule = fieldValue as number < validationList.validatorList[fieldName][ruleName];
+                        if (fieldHasValue === true) fieldPassesRule = fieldValue < validationList.validatorList[fieldName][ruleName];
                         break;
                     case "greaterThan":
-                        if (fieldHasValue === true) fieldPassesRule = fieldValue as number > validationList.validatorList[fieldName][ruleName];
+                        if (fieldHasValue === true) fieldPassesRule = fieldValue > validationList.validatorList[fieldName][ruleName];
                         break;
                     case "lessThanOrEqualTo":
-                        if (fieldHasValue === true) fieldPassesRule = fieldValue as number <= validationList.validatorList[fieldName][ruleName];
+                        if (fieldHasValue === true) fieldPassesRule = fieldValue <= validationList.validatorList[fieldName][ruleName];
                         break;
                     case "greaterThanOrEqualTo":
-                        if (fieldHasValue === true) fieldPassesRule = fieldValue as number >= validationList.validatorList[fieldName][ruleName];
+                        if (fieldHasValue === true) fieldPassesRule = fieldValue >= validationList.validatorList[fieldName][ruleName];
                         break;
                     case "minLength":
                         if (fieldHasValue === true) fieldPassesRule = (fieldValue as string).length >= (validationList.validatorList[fieldName][ruleName] as number);
@@ -187,7 +187,7 @@ $(function () {
     };
 
     // Add the button-click event that populates the form
-    $("#populateFormButton").click(function () {
+    $("#populateFormButton").on("click", function () {
         resetResult();
         $.getJSON("api/Person/1", (json) => populateFormFromJson({ json }));
     });
